Refactored the articles and toolbar JS to isolate WebExtension vs JPM api calls.
Tested on Firefox, WebExtensions still works.
WIP JPM still untested.
