+++ data/settings/index.html
@@ -39,6 +39,7 @@
   <script type="text/javascript" src="../../lib/spin.js"></script>
   <script type="text/javascript" src="tablesort.js"></script>
   <script type="text/javascript" src="articles-table-chrome.js"></script>
+  <script type="text/javascript" src="articles-table-content.js"></script>
   <script type="text/javascript" src="articles-table-lib.js"></script>
 
   <script type="text/javascript" src="../../lib/DataTables-1.10.7/media/js/jquery.dataTables.js"></script>
