+++ manifest.json-chrome
@@ -18,6 +18,7 @@
       "js": [
         "lib/jquery-1.11.2.min.js", "src/ao3lib.js", 
         "src/toolbar-ao3.js", "src/toolbar-chrome.js",
+        "src/toolbar-content.js",
         "src/article.js", "src/images.js"
       ],
       "all_frames": true
